+++ pkg/artifactory/resource_artifactory_local_repository_test.go
@@ -11,7 +11,7 @@ import (
 	"net/http"
 )
 
-const localRepository_basic = `
+const localRepositoryBasic = `
 resource "artifactory_local_repository" "terraform-local-test-repo-basic" {
 	key 	     = "terraform-local-test-repo-basic"
 	package_type = "docker"
@@ -24,7 +24,7 @@ func TestAccLocalRepository_basic(t *testing.T) {
 		Providers:    testAccProviders,
 		Steps: []resource.TestStep{
 			{
-				Config: localRepository_basic,
+				Config: localRepositoryBasic,
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("artifactory_local_repository.terraform-local-test-repo-basic", "key", "terraform-local-test-repo-basic"),
 					resource.TestCheckResourceAttr("artifactory_local_repository.terraform-local-test-repo-basic", "package_type", "docker"),
@@ -34,7 +34,7 @@ func TestAccLocalRepository_basic(t *testing.T) {
 	})
 }
 
-const localRepositoryConfig_full = `
+const localRepositoryConfigFull = `
 resource "artifactory_local_repository" "terraform-local-test-repo-full" {
     key                             = "terraform-local-test-repo-full"
     package_type                    = "npm"
@@ -66,7 +66,7 @@ func TestAccLocalRepository_full(t *testing.T) {
 		CheckDestroy: resourceLocalRepositoryCheckDestroy("artifactory_local_repository.terraform-local-test-repo-full"),
 		Steps: []resource.TestStep{
 			{
-				Config: localRepositoryConfig_full,
+				Config: localRepositoryConfigFull,
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("artifactory_local_repository.terraform-local-test-repo-full", "key", "terraform-local-test-repo-full"),
 					resource.TestCheckResourceAttr("artifactory_local_repository.terraform-local-test-repo-full", "package_type", "npm"),