+++ pkg/artifactory/resource_artifactory_replication_config_test.go
@@ -1,30 +1,24 @@
 package artifactory
 
 import (
-	"fmt"
-	"os"
-	"testing"
-
 	"context"
+	"fmt"
 	"github.com/atlassian/go-artifactory/pkg/artifactory"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 	"net/http"
+	"os"
+	"testing"
 )
 
 const replicationConfigTemplate = `
-resource "artifactory_local_repository" "provider_test_source" {
-	key = "provider_test_source"
+resource "artifactory_local_repository" "lib-local" {
+	key = "lib-local"
 	package_type = "maven"
 }
 
-resource "artifactory_local_repository" "provider_test_dest" {
-	key = "provider_test_dest"
-	package_type = "maven"
-}
-
-resource "artifactory_replication_config" "foo-rep" {
-	repo_key = "${artifactory_local_repository.provider_test_source.key}"
+resource "artifactory_replication_config" "lib-local" {
+	repo_key = "${artifactory_local_repository.lib-local.key}"
 	cron_exp = "0 0 * * * ?"
 	enable_event_replication = true
 	
@@ -40,18 +34,22 @@ resource "artifactory_replication_config" "foo-rep" {
 
 func TestAccReplication_full(t *testing.T) {
 	resource.Test(t, resource.TestCase{
-		PreCheck:     func() {},
-		CheckDestroy: testAccCheckReplicationDestroy("artifactory_replication_config.foo"),
+		CheckDestroy: testAccCheckReplicationDestroy("artifactory_replication_config.lib-local"),
 		Providers:    testAccProviders,
 
 		Steps: []resource.TestStep{
 			{
-				Config: fmt.Sprintf(replicationConfigTemplate, os.Getenv("ARTIFACTORY_URL"), os.Getenv("ARTIFACTORY_USERNAME"), os.Getenv("ARTIFACTORY_PASSWORD")),
+				Config: fmt.Sprintf(
+					replicationConfigTemplate,
+					os.Getenv("ARTIFACTORY_URL"),
+					os.Getenv("ARTIFACTORY_USERNAME"),
+					os.Getenv("ARTIFACTORY_PASSWORD"),
+				),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr("artifactory_replication_config.foo-rep", "repo_key", "provider_test_source"),
-					resource.TestCheckResourceAttr("artifactory_replication_config.foo-rep", "cron_exp", "0 0 * * * ?"),
-					resource.TestCheckResourceAttr("artifactory_replication_config.foo-rep", "enable_event_replication", "true"),
-					resource.TestCheckResourceAttr("artifactory_replication_config.foo-rep", "replications.#", "1"),
+					resource.TestCheckResourceAttr("artifactory_replication_config.lib-local", "repo_key", "lib-local"),
+					resource.TestCheckResourceAttr("artifactory_replication_config.lib-local", "cron_exp", "0 0 * * * ?"),
+					resource.TestCheckResourceAttr("artifactory_replication_config.lib-local", "enable_event_replication", "true"),
+					resource.TestCheckResourceAttr("artifactory_replication_config.lib-local", "replications.#", "1"),
 				),
 			},
 		},
@@ -68,7 +66,7 @@ func testAccCheckReplicationDestroy(id string) func(*terraform.State) error {
 
 		replica, resp, err := client.Artifacts.GetRepositoryReplicationConfig(context.Background(), rs.Primary.ID)
 
-		if resp.StatusCode == http.StatusNotFound {
+		if resp.StatusCode == http.StatusNotFound || resp.StatusCode == http.StatusBadRequest {
 			return nil
 		} else if err != nil {
 			return fmt.Errorf("error: Request failed: %s", err.Error())